requirements.txt
pytest
python-dotenv
beautifulsoup4
aiohttp
